requirements.txt
asgiref==3.5.2
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
mysqlclient==2.1.1
Pygments==2.12.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
zipp==3.8.1
